requirements.txt
Django==2.1.1
django-crispy-forms==1.9.0
django-phonenumber-field==4.0.0
face-recognition==1.3.0
mysqlclient==1.4.6
phonenumbers==8.11.5
Pillow==7.0.0
requests==2.20.1